move ZeroXPrefixedBytes out of module jsonrpc
it really doesn't have anything to do with JSON-RPC specifically
