+++ crates/rethnet_eth/src/remote/jsonrpc.rs
@@ -113,54 +113,3 @@ impl<'a> serde::de::Visitor<'a> for VersionVisitor {
         }
     }
 }
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct ZeroXPrefixedBytes {
-    pub inner: bytes::Bytes,
-}
-
-impl<'a> Deserialize<'a> for ZeroXPrefixedBytes {
-    fn deserialize<D>(deserializer: D) -> Result<ZeroXPrefixedBytes, D::Error>
-    where
-        D: serde::Deserializer<'a>,
-    {
-        struct ZeroXPrefixedBytesVisitor;
-        impl<'a> serde::de::Visitor<'a> for ZeroXPrefixedBytesVisitor {
-            type Value = ZeroXPrefixedBytes;
-
-            fn expecting(&self, formatter: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-                formatter.write_str("a 0x-prefixed string of hex digits")
-            }
-
-            fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                if &value[0..1] == "0x" {
-                    Err(serde::de::Error::custom(
-                        "string does not have a '0x' prefix",
-                    ))
-                } else {
-                    Ok(ZeroXPrefixedBytes {
-                        inner: bytes::Bytes::from(
-                            hex::decode(&value[2..]).unwrap_or_else(|_| {
-                                panic!("failed to decode hex string \"{value}\"")
-                            }),
-                        ),
-                    })
-                }
-            }
-        }
-
-        deserializer.deserialize_identifier(ZeroXPrefixedBytesVisitor)
-    }
-}
-
-impl Serialize for ZeroXPrefixedBytes {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.serialize_str(&format!("0x{}", hex::encode(self.inner.clone()),))
-    }
-}
